refactor: restdocs 관련 util 리팩토링
