--- lovebird-api/src/test/kotlin/com/lovebird/api/restdoc/config/RestDocsConfig.kt
@@ -1,19 +1,0 @@
-package com.lovebird.api.restdoc.config
-
-import org.springframework.boot.test.autoconfigure.restdocs.RestDocsMockMvcConfigurationCustomizer
-import org.springframework.boot.test.context.TestConfiguration
-import org.springframework.context.annotation.Bean
-import org.springframework.restdocs.operation.preprocess.Preprocessors.prettyPrint
-
-@TestConfiguration
-class RestDocsConfig {
-
-	@Bean
-	fun restDocsMockMvcConfigurationCustomizer(): RestDocsMockMvcConfigurationCustomizer {
-		return RestDocsMockMvcConfigurationCustomizer { configurer ->
-			configurer.operationPreprocessors()
-				.withRequestDefaults(prettyPrint())
-				.withResponseDefaults(prettyPrint())
-		}
-	}
-}
